refactor/docs: clean up
