+++ src/BibInput.js
@@ -1,13 +1,18 @@
 import React, { useState, useEffect } from 'react';
+import { requestData } from './data/api';
 
+/**
+ * Component for handling user input for year and bib number
+ * @param {function} onObjectReceived called when response from MyRaceResult is returned
+ * @returns year and bib input controls
+ */
 const BibInput = ({ onObjectReceived }) => {
-
-  const [inputValue, setInputValue] = useState(0);
+  const [inputValue, setInputValue] = useState(0); // Keeps track of input to bib number field
 
   const handleInputChange = (e) => {
     setInputValue(e.target.value);
   };
-  const [selectedValue, setSelectedValue] = useState("2024"); // State to keep track of selected value
+  const [selectedValue, setSelectedValue] = useState("2024"); // State to keep track of years dropdown
 
   const handleChange = (event) => {
     setSelectedValue(event.target.value); // Update selected value when dropdown value changes
@@ -22,8 +27,9 @@ const BibInput = ({ onObjectReceived }) => {
     if (yearParam && ['2022', '2023', '2024'].includes(yearParam)) {
       setSelectedValue(yearParam); // Assign the year parameter to loadedYear state
     }
+    // Check if the bib parameter exists in the URL
     if (bibParam) {
-      setInputValue(bibParam);
+      setInputValue(bibParam); // Assign the bib parameter to inputValue
     }
   }, []); // Only run this effect once, when the component mounts
 
@@ -31,37 +37,23 @@ const BibInput = ({ onObjectReceived }) => {
     if (inputValue) {
       handleButtonClick();
     }
-  }, [inputValue]);
+  }, [inputValue]); // Run every time input changes
 
   const handleButtonClick = () => {
-    const xhr = new XMLHttpRequest();
     // xhr.open('GET', 'https://my4.raceresult.com/192607/RRPublish/data/list?key=9d484a9a9259ff0ae1a4a8570861bc3b&listname=Online%7CLap%20Details&page=live&contest=0&r=bib2&bib=' + inputValue); // 2022
     // xhr.open('GET', 'https://my4.raceresult.com/204047/RRPublish/data/list?key=b02d8bcb6d81d09372a43de65f7f7d48&listname=Online%7CLap%20Details&page=live&contest=0&r=bib2&bib=' + inputValue); // 2023
     // xhr.open('GET', 'https://my1.raceresult.com/259072/RRPublish/data/list?key=eca2e3d1510caee33b7710a250a6f2c1&listname=Online%7CLap%20Details&page=live&contest=0&r=bib2&bib=' + inputValue); // 2024
     // xhr.open('GET', 'https://raw.githubusercontent.com/PeterChu3/jsonHosting/main/12.json'); // Dummy data
 
     if (selectedValue === "2024") {
-      xhr.open('GET', 'https://my1.raceresult.com/259072/RRPublish/data/list?key=eca2e3d1510caee33b7710a250a6f2c1&listname=Online%7CLap%20Details&page=live&contest=0&r=bib2&bib=' + inputValue); // 2024
-    } else if (selectedValue === "2023") {
-      xhr.open('GET', 'https://my4.raceresult.com/204047/RRPublish/data/list?key=b02d8bcb6d81d09372a43de65f7f7d48&listname=Online%7CLap%20Details&page=live&contest=0&r=bib2&bib=' + inputValue); // 2023
-    } else if (selectedValue === "2022") {
-      xhr.open('GET', 'https://my4.raceresult.com/192607/RRPublish/data/list?key=9d484a9a9259ff0ae1a4a8570861bc3b&listname=Online%7CLap%20Details&page=live&contest=0&r=bib2&bib=' + inputValue); // 2022
-    }
-    xhr.onload = function () {
-      if (xhr.status === 200) {
-        onObjectReceived(JSON.parse(xhr.responseText));
-
-      }
-    };
-    xhr.send();
-
-
-    // if (data != null) {
-    //   console.log(data.data);
-    // }
+        // By extracting this function, you can now use it in other parts of the app and reliably test it if you ever wanted to add tests
+        // You can also pass a different callback function parameter if you wanted to do something different with the response
+        requestData(selectedValue, inputValue, onObjectReceived);
   };
+}
 
-
+  // SUGGEST: Using a CSS library would make styling something like this pretty simple
+  // Recommendations would be TailwindCSS (my favorite) or Bootstrap (the most popular)
   return (
     <div>
       <select id="dropdown" value={selectedValue} onChange={handleChange}>
